pyproject.toml
[tool.poetry]
name = "f1-pit-optimizer"
version = "0.1.0"
description = ""
authors = ["Avi <Aviozawa@gmail.com>"]
readme = "README.md"
packages = [{include = "f1_pit_optimizer", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11"
requests = ">=2.32.5,<3.0.0"
pandas = ">=2.3.2,<3.0.0"
matplotlib = ">=3.10.6,<4.0.0"
fastf1 = ">=3.6.1,<4.0.0"
tabulate = ">=0.9.0,<0.10.0"
scikit-learn = ">=1.7.2,<2.0.0"
seaborn = "^0.13.2"
python-dotenv = "^1.1.1"
streamlit = "^1.49.1"
google-genai = "^1.36.0"
[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"